move methods from lisp.rs to related files
