+++ rust_src/src/lisp.rs
@@ -19,27 +19,18 @@ use remacs_sys::{pvec_type, EmacsDouble, EmacsInt, EmacsUint, EqualKind, Fcons, 
                  MOST_NEGATIVE_FIXNUM, MOST_POSITIVE_FIXNUM, USE_LSB_TAG, VALMASK};
 use remacs_sys::{Lisp_Cons, Lisp_Float, Lisp_Misc_Any, Lisp_Misc_Type, Lisp_Subr, Lisp_Symbol,
                  Lisp_Type};
-use remacs_sys::{Qarrayp, Qautoload, Qbufferp, Qchar_table_p, Qcharacterp, Qconsp, Qfloatp,
-                 Qframe_live_p, Qframep, Qhash_table_p, Qinteger_or_marker_p, Qintegerp, Qlistp,
-                 Qmarkerp, Qnil, Qnumber_or_marker_p, Qnumberp, Qoverlayp, Qplistp, Qprocessp,
-                 Qstringp, Qsubrp, Qsymbolp, Qt, Qthreadp, Qunbound, Qvectorp, Qwholenump,
-                 Qwindow_live_p, Qwindow_valid_p, Qwindowp, Vbuffer_alist};
+use remacs_sys::{Qarrayp, Qautoload, Qchar_table_p, Qcharacterp, Qconsp, Qfloatp,
+                 Qinteger_or_marker_p, Qintegerp, Qlistp, Qmarkerp, Qnil, Qnumber_or_marker_p,
+                 Qnumberp, Qplistp, Qstringp, Qsubrp, Qsymbolp, Qt, Qunbound, Qvectorp,
+                 Qwholenump, Vbuffer_alist};
 
-use buffers::{LispBufferRef, LispOverlayRef};
 use chartable::{LispCharTableRef, LispSubCharTableAsciiRef, LispSubCharTableRef};
 use eval::FUNCTIONP;
 use fonts::LispFontRef;
-use frames::LispFrameRef;
-use hashtable::LispHashTableRef;
 use lists::{circular_list, list};
-use marker::LispMarkerRef;
 use multibyte::{Codepoint, LispStringRef, MAX_CHAR};
 use obarray::{check_obarray, LispObarrayRef};
-use process::LispProcessRef;
-use symbols::LispSymbolRef;
-use threads::ThreadStateRef;
 use vectors::{LispBoolVecRef, LispVectorRef, LispVectorlikeRef};
-use windows::LispWindowRef;
 
 // TODO: tweak Makefile to rebuild C files if this changes.
 
@@ -223,77 +214,6 @@ impl From<LispObject> for Option<LispObarrayRef> {
     }
 }
 
-// Symbol support (LispType == Lisp_Symbol == 0)
-impl LispObject {
-    #[inline]
-    pub fn is_symbol(self) -> bool {
-        self.get_type() == Lisp_Type::Lisp_Symbol
-    }
-
-    #[inline]
-    pub fn as_symbol(self) -> Option<LispSymbolRef> {
-        if self.is_symbol() {
-            Some(LispSymbolRef::new(
-                self.symbol_ptr_value() as *mut Lisp_Symbol
-            ))
-        } else {
-            None
-        }
-    }
-
-    #[inline]
-    pub fn as_symbol_or_error(self) -> LispSymbolRef {
-        if let Some(sym) = self.as_symbol() {
-            sym
-        } else {
-            wrong_type!(Qsymbolp, self)
-        }
-    }
-
-    #[inline]
-    pub fn symbol_or_string_as_string(self) -> LispStringRef {
-        match self.as_symbol() {
-            Some(sym) => sym
-                .symbol_name()
-                .as_string()
-                .expect("Expected a symbol name?"),
-            None => self.as_string_or_error(),
-        }
-    }
-
-    #[inline]
-    fn symbol_ptr_value(self) -> EmacsInt {
-        let ptr_value = if USE_LSB_TAG {
-            self.to_C()
-        } else {
-            self.get_untaggedptr() as EmacsInt
-        };
-
-        let lispsym_offset = unsafe { &lispsym as *const _ as EmacsInt };
-        ptr_value + lispsym_offset
-    }
-}
-
-impl From<LispObject> for LispSymbolRef {
-    #[inline]
-    fn from(o: LispObject) -> Self {
-        o.as_symbol_or_error()
-    }
-}
-
-impl From<LispSymbolRef> for LispObject {
-    #[inline]
-    fn from(s: LispSymbolRef) -> Self {
-        s.as_lisp_obj()
-    }
-}
-
-impl From<LispObject> for Option<LispSymbolRef> {
-    fn from(o: LispObject) -> Option<LispSymbolRef> {
-        o.as_symbol()
-    }
-}
-
 // Misc support (LispType == Lisp_Misc == 1)
 
 // Lisp_Misc is a union. Now we don't really care about its variants except the
@@ -416,127 +336,6 @@ impl LispObject {
     }
 }
 
-// Fixnum(Integer) support (LispType == Lisp_Int0 | Lisp_Int1 == 2 | 6(LSB) )
-
-/// Fixnums are inline integers that fit directly into Lisp's tagged word.
-/// There's two `LispType` variants to provide an extra bit.
-
-/// Natnums(natural number) are the non-negative fixnums.
-/// There were special branches in the original code for better performance.
-/// However they are unified into the fixnum logic under LSB mode.
-/// TODO: Recheck these logic in original C code.
-
-impl LispObject {
-    #[inline]
-    pub fn from_fixnum(n: EmacsInt) -> LispObject {
-        debug_assert!(MOST_NEGATIVE_FIXNUM <= n && n <= MOST_POSITIVE_FIXNUM);
-        Self::from_fixnum_truncated(n)
-    }
-
-    #[inline]
-    pub fn from_fixnum_truncated(n: EmacsInt) -> LispObject {
-        let o = if USE_LSB_TAG {
-            (n << Lisp_Bits::INTTYPEBITS) as EmacsUint + Lisp_Type::Lisp_Int0 as EmacsUint
-        } else {
-            (n & INTMASK) as EmacsUint + ((Lisp_Type::Lisp_Int0 as EmacsUint) << Lisp_Bits::VALBITS)
-        };
-        LispObject::from_C(o as EmacsInt)
-    }
-
-    /// Convert a positive integer into its LispObject representation.
-    ///
-    /// This is also the function to use when translating `XSETFASTINT`
-    /// from Emacs C.
-    // TODO: the C claims that make_natnum is faster, but it does the same
-    // thing as make_number when USE_LSB_TAG is 1, which it is for us. We
-    // should remove this in favour of make_number.
-    #[inline]
-    pub fn from_natnum(n: EmacsUint) -> LispObject {
-        debug_assert!(n <= (MOST_POSITIVE_FIXNUM as EmacsUint));
-        LispObject::from_fixnum_truncated(n as EmacsInt)
-    }
-
-    #[inline]
-    pub fn int_or_float_from_fixnum(n: EmacsInt) -> LispObject {
-        if n < MOST_NEGATIVE_FIXNUM || n > MOST_POSITIVE_FIXNUM {
-            Self::from_float(n as f64)
-        } else {
-            Self::from_fixnum(n)
-        }
-    }
-
-    #[inline]
-    pub fn fixnum_overflow(n: EmacsInt) -> bool {
-        n < MOST_NEGATIVE_FIXNUM || n > MOST_POSITIVE_FIXNUM
-    }
-
-    #[inline]
-    pub unsafe fn to_fixnum_unchecked(self) -> EmacsInt {
-        let raw = self.to_C();
-        if !USE_LSB_TAG {
-            raw & INTMASK
-        } else {
-            raw >> Lisp_Bits::INTTYPEBITS
-        }
-    }
-
-    #[inline]
-    pub fn is_fixnum(self) -> bool {
-        let ty = self.get_type();
-        (ty as u8 & ((Lisp_Type::Lisp_Int0 as u8) | !(Lisp_Type::Lisp_Int1 as u8)))
-            == Lisp_Type::Lisp_Int0 as u8
-    }
-
-    #[inline]
-    pub fn as_fixnum(self) -> Option<EmacsInt> {
-        if self.is_fixnum() {
-            Some(unsafe { self.to_fixnum_unchecked() })
-        } else {
-            None
-        }
-    }
-
-    #[inline]
-    pub fn as_fixnum_or_error(self) -> EmacsInt {
-        if self.is_fixnum() {
-            unsafe { self.to_fixnum_unchecked() }
-        } else {
-            wrong_type!(Qintegerp, self)
-        }
-    }
-
-    #[inline]
-    pub fn as_fixnum_coerce_marker_or_error(self) -> EmacsInt {
-        if let Some(n) = self.as_fixnum() {
-            n
-        } else if let Some(m) = self.as_marker() {
-            m.charpos_or_error() as EmacsInt
-        } else {
-            wrong_type!(Qinteger_or_marker_p, self);
-        }
-    }
-
-    /// TODO: Bignum support? (Current Emacs doesn't have it)
-    #[inline]
-    pub fn is_integer(self) -> bool {
-        self.is_fixnum()
-    }
-
-    #[inline]
-    pub fn is_natnum(self) -> bool {
-        self.as_fixnum().map_or(false, |i| i >= 0)
-    }
-
-    #[inline]
-    pub fn as_natnum_or_error(self) -> EmacsUint {
-        if self.is_natnum() {
-            unsafe { self.to_fixnum_unchecked() as EmacsUint }
-        } else {
-            wrong_type!(Qwholenump, self)
-        }
-    }
-}
-
 impl From<LispObject> for EmacsInt {
     #[inline]
     fn from(o: LispObject) -> Self {
@@ -634,86 +433,7 @@ impl From<u8> for LispObject {
     }
 }
 
-// Vectorlike support (LispType == 5)
-
 impl LispObject {
-    #[inline]
-    pub fn is_vectorlike(self) -> bool {
-        self.get_type() == Lisp_Type::Lisp_Vectorlike
-    }
-
-    #[inline]
-    pub fn is_vector(self) -> bool {
-        self.as_vectorlike().map_or(false, |v| v.is_vector())
-    }
-
-    #[inline]
-    pub fn as_vectorlike(self) -> Option<LispVectorlikeRef> {
-        if self.is_vectorlike() {
-            Some(LispVectorlikeRef::new(
-                self.get_untaggedptr() as *mut remacs_sys::Lisp_Vectorlike
-            ))
-        } else {
-            None
-        }
-    }
-
-    /*
-    #[inline]
-    pub fn as_vectorlike_or_error(self) -> LispVectorlikeRef {
-        if self.is_vectorlike() {
-            LispVectorlikeRef::new(unsafe { mem::transmute(self.get_untaggedptr()) })
-        } else {
-            wrong_type!(Qvectorp, self)
-        }
-    }
-    */
-
-    pub unsafe fn as_vectorlike_unchecked(self) -> LispVectorlikeRef {
-        LispVectorlikeRef::new(self.get_untaggedptr() as *mut remacs_sys::Lisp_Vectorlike)
-    }
-
-    pub fn as_vector(self) -> Option<LispVectorRef> {
-        self.as_vectorlike().and_then(|v| v.as_vector())
-    }
-
-    pub fn as_vector_or_error(self) -> LispVectorRef {
-        self.as_vector()
-            .unwrap_or_else(|| wrong_type!(Qvectorp, self))
-    }
-
-    pub unsafe fn as_vector_unchecked(self) -> LispVectorRef {
-        self.as_vectorlike_unchecked().as_vector_unchecked()
-    }
-
-    pub fn as_vector_or_string_length(self) -> isize {
-        if let Some(s) = self.as_string() {
-            return s.len_chars();
-        } else if let Some(vl) = self.as_vectorlike() {
-            if let Some(v) = vl.as_vector() {
-                return v.len() as isize;
-            }
-        };
-
-        wrong_type!(Qarrayp, self);
-    }
-}
-
-impl LispObject {
-    pub fn is_thread(self) -> bool {
-        self.as_vectorlike()
-            .map_or(false, |v| v.is_pseudovector(pvec_type::PVEC_THREAD))
-    }
-
-    pub fn as_thread(self) -> Option<ThreadStateRef> {
-        self.as_vectorlike().and_then(|v| v.as_thread())
-    }
-
-    pub fn as_thread_or_error(self) -> ThreadStateRef {
-        self.as_thread()
-            .unwrap_or_else(|| wrong_type!(Qthreadp, self))
-    }
-
     pub fn is_mutex(self) -> bool {
         self.as_vectorlike()
             .map_or(false, |v| v.is_pseudovector(pvec_type::PVEC_MUTEX))
@@ -746,25 +466,6 @@ impl LispObject {
 
     pub fn as_subr_or_error(self) -> LispSubrRef {
         self.as_subr().unwrap_or_else(|| wrong_type!(Qsubrp, self))
-    }
-
-    pub fn is_buffer(self) -> bool {
-        self.as_vectorlike()
-            .map_or(false, |v| v.is_pseudovector(pvec_type::PVEC_BUFFER))
-    }
-
-    pub fn as_buffer(self) -> Option<LispBufferRef> {
-        self.as_vectorlike().and_then(|v| v.as_buffer())
-    }
-
-    pub fn as_live_buffer(self) -> Option<LispBufferRef> {
-        self.as_buffer()
-            .and_then(|b| if b.is_live() { Some(b) } else { None })
-    }
-
-    pub fn as_buffer_or_error(self) -> LispBufferRef {
-        self.as_buffer()
-            .unwrap_or_else(|| wrong_type!(Qbufferp, self))
     }
 
     pub fn is_char_table(self) -> bool {
@@ -810,102 +511,6 @@ impl LispObject {
         self.is_cons() || self.is_nil() || self.is_array()
     }
 
-    /*
-    pub fn is_window_configuration(self) -> bool {
-        self.as_vectorlike().map_or(
-            false,
-            |v| v.is_pseudovector(pvec_type::PVEC_WINDOW_CONFIGURATION),
-        )
-    }
-    */
-
-    pub fn is_process(self) -> bool {
-        self.as_vectorlike()
-            .map_or(false, |v| v.is_pseudovector(pvec_type::PVEC_PROCESS))
-    }
-
-    pub fn as_process(self) -> Option<LispProcessRef> {
-        self.as_vectorlike().and_then(|v| v.as_process())
-    }
-
-    pub fn as_process_or_error(self) -> LispProcessRef {
-        self.as_process()
-            .unwrap_or_else(|| wrong_type!(Qprocessp, self))
-    }
-
-    pub fn is_window(self) -> bool {
-        self.as_vectorlike()
-            .map_or(false, |v| v.is_pseudovector(pvec_type::PVEC_WINDOW))
-    }
-
-    pub fn as_window(self) -> Option<LispWindowRef> {
-        self.as_vectorlike().and_then(|v| v.as_window())
-    }
-
-    pub fn as_window_or_error(self) -> LispWindowRef {
-        self.as_window()
-            .unwrap_or_else(|| wrong_type!(Qwindowp, self))
-    }
-
-    pub fn as_minibuffer_or_error(self) -> LispWindowRef {
-        let w = self
-            .as_window()
-            .unwrap_or_else(|| wrong_type!(Qwindowp, self));
-        if !w.is_minibuffer() {
-            error!("Window is not a minibuffer window");
-        }
-        w
-    }
-
-    pub fn as_live_window(self) -> Option<LispWindowRef> {
-        self.as_window()
-            .and_then(|w| if w.is_live() { Some(w) } else { None })
-    }
-
-    pub fn as_live_window_or_error(self) -> LispWindowRef {
-        self.as_live_window()
-            .unwrap_or_else(|| wrong_type!(Qwindow_live_p, self))
-    }
-
-    pub fn as_valid_window(self) -> Option<LispWindowRef> {
-        self.as_window()
-            .and_then(|w| if w.is_valid() { Some(w) } else { None })
-    }
-
-    pub fn as_valid_window_or_error(self) -> LispWindowRef {
-        self.as_valid_window()
-            .unwrap_or_else(|| wrong_type!(Qwindow_valid_p, self))
-    }
-
-    pub fn is_frame(self) -> bool {
-        self.as_vectorlike()
-            .map_or(false, |v| v.is_pseudovector(pvec_type::PVEC_FRAME))
-    }
-
-    pub fn as_frame(self) -> Option<LispFrameRef> {
-        self.as_vectorlike().and_then(|v| v.as_frame())
-    }
-
-    pub fn as_frame_or_error(self) -> LispFrameRef {
-        self.as_frame()
-            .unwrap_or_else(|| wrong_type!(Qframep, self))
-    }
-
-    pub fn as_live_frame(self) -> Option<LispFrameRef> {
-        self.as_frame()
-            .and_then(|f| if f.is_live() { Some(f) } else { None })
-    }
-
-    pub fn as_live_frame_or_error(self) -> LispFrameRef {
-        self.as_live_frame()
-            .unwrap_or_else(|| wrong_type!(Qframe_live_p, self))
-    }
-
-    pub fn is_hash_table(self) -> bool {
-        self.as_vectorlike()
-            .map_or(false, |v| v.is_pseudovector(pvec_type::PVEC_HASH_TABLE))
-    }
-
     pub fn is_font(self) -> bool {
         self.as_vectorlike()
             .map_or(false, |v| v.is_pseudovector(pvec_type::PVEC_FONT))
@@ -945,44 +550,6 @@ impl LispObject {
     }
 }
 
-impl From<LispObject> for LispWindowRef {
-    fn from(o: LispObject) -> Self {
-        o.as_window_or_error()
-    }
-}
-
-impl From<LispWindowRef> for LispObject {
-    fn from(w: LispWindowRef) -> Self {
-        w.as_lisp_obj()
-    }
-}
-
-impl From<LispObject> for Option<LispWindowRef> {
-    #[inline]
-    fn from(o: LispObject) -> Self {
-        o.as_window()
-    }
-}
-
-impl From<LispObject> for LispFrameRef {
-    fn from(o: LispObject) -> Self {
-        o.as_frame_or_error()
-    }
-}
-
-impl From<LispFrameRef> for LispObject {
-    fn from(f: LispFrameRef) -> Self {
-        f.as_lisp_obj()
-    }
-}
-
-impl From<LispObject> for Option<LispFrameRef> {
-    #[inline]
-    fn from(o: LispObject) -> Self {
-        o.as_frame()
-    }
-}
-
 impl From<LispObject> for LispCharTableRef {
     fn from(o: LispObject) -> Self {
         o.as_char_table_or_error()
@@ -1005,169 +572,6 @@ impl From<LispObject> for LispSubrRef {
     #[inline]
     fn from(o: LispObject) -> Self {
         o.as_subr_or_error()
-    }
-}
-
-impl From<LispObject> for LispBufferRef {
-    #[inline]
-    fn from(o: LispObject) -> Self {
-        o.as_buffer_or_error()
-    }
-}
-
-impl From<LispBufferRef> for LispObject {
-    fn from(b: LispBufferRef) -> Self {
-        b.as_lisp_obj()
-    }
-}
-
-impl From<LispObject> for Option<LispBufferRef> {
-    #[inline]
-    fn from(o: LispObject) -> Self {
-        o.as_buffer()
-    }
-}
-
-impl From<LispObject> for ThreadStateRef {
-    #[inline]
-    fn from(o: LispObject) -> Self {
-        o.as_thread_or_error()
-    }
-}
-
-impl LispObject {
-    pub fn as_hash_table_or_error(self) -> LispHashTableRef {
-        if self.is_hash_table() {
-            LispHashTableRef::new(self.get_untaggedptr() as *mut remacs_sys::Lisp_Hash_Table)
-        } else {
-            wrong_type!(Qhash_table_p, self);
-        }
-    }
-
-    pub fn from_hash_table(hashtable: LispHashTableRef) -> LispObject {
-        let object = LispObject::tag_ptr(hashtable, Lisp_Type::Lisp_Vectorlike);
-        debug_assert!(
-            object.is_vectorlike() && object.get_untaggedptr() == hashtable.as_ptr() as *mut c_void
-        );
-
-        debug_assert!(object.is_hash_table());
-        object
-    }
-}
-
-impl From<LispObject> for LispHashTableRef {
-    #[inline]
-    fn from(o: LispObject) -> Self {
-        o.as_hash_table_or_error()
-    }
-}
-
-impl From<LispHashTableRef> for LispObject {
-    #[inline]
-    fn from(h: LispHashTableRef) -> Self {
-        LispObject::from_hash_table(h)
-    }
-}
-
-// Cons support (LispType == 6 | 3)
-
-/// From `FOR_EACH_TAIL_INTERNAL` in `lisp.h`
-pub struct TailsIter {
-    list: LispObject,
-    tail: LispObject,
-    tortoise: LispObject,
-    errsym: Option<LispObject>,
-    max: isize,
-    n: isize,
-    q: u16,
-}
-
-impl TailsIter {
-    fn new(list: LispObject, errsym: Option<LispObject>) -> Self {
-        Self {
-            list,
-            tail: list,
-            tortoise: list,
-            errsym,
-            max: 2,
-            n: 0,
-            q: 2,
-        }
-    }
-
-    pub fn rest(&self) -> LispObject {
-        // This is kind of like Peekable but even when None is returned there
-        // might still be a valid item in self.tail.
-        self.tail
-    }
-
-    fn circular(&self) -> Option<LispCons> {
-        if self.errsym.is_some() {
-            circular_list(self.tail);
-        } else {
-            None
-        }
-    }
-}
-
-impl Iterator for TailsIter {
-    type Item = LispCons;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        match self.tail.as_cons() {
-            None => {
-                if self.errsym.is_some() && self.tail.is_not_nil() {
-                    wrong_type!(self.errsym.unwrap(), self.list)
-                }
-                None
-            }
-            Some(tail_cons) => {
-                self.tail = tail_cons.cdr();
-                self.q = self.q.wrapping_sub(1);
-                if self.q != 0 {
-                    if self.tail == self.tortoise {
-                        return self.circular();
-                    }
-                } else {
-                    self.n = self.n.wrapping_sub(1);
-                    if self.n > 0 {
-                        if self.tail == self.tortoise {
-                            return self.circular();
-                        }
-                    } else {
-                        self.max <<= 1;
-                        self.q = self.max as u16;
-                        self.n = self.max >> 16;
-                        self.tortoise = self.tail;
-                    }
-                }
-                Some(tail_cons)
-            }
-        }
-    }
-}
-
-pub struct CarIter {
-    tails: TailsIter,
-}
-
-impl CarIter {
-    pub fn new(list: LispObject, errsym: Option<LispObject>) -> Self {
-        Self {
-            tails: TailsIter::new(list, errsym),
-        }
-    }
-
-    pub fn rest(&self) -> LispObject {
-        self.tails.tail
-    }
-}
-
-impl Iterator for CarIter {
-    type Item = LispObject;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.tails.next().map(|c| c.car())
     }
 }
 
@@ -1204,356 +608,15 @@ macro_rules! impl_alistval_iter {
 
 impl_alistval_iter! {LiveBufferIter, LispBufferRef, unsafe { Vbuffer_alist }}
 
-impl LispObject {
-    #[inline]
-    pub fn cons(car: LispObject, cdr: LispObject) -> Self {
-        unsafe { Fcons(car, cdr) }
-    }
-
-    #[inline]
-    pub fn is_cons(self) -> bool {
-        self.get_type() == Lisp_Type::Lisp_Cons
-    }
-
-    #[inline]
-    pub fn as_cons(self) -> Option<LispCons> {
-        if self.is_cons() {
-            Some(LispCons(self))
-        } else {
-            None
-        }
-    }
-
-    #[inline]
-    pub fn as_cons_or_error(self) -> LispCons {
-        if self.is_cons() {
-            LispCons(self)
-        } else {
-            wrong_type!(Qconsp, self)
-        }
-    }
-
-    #[inline]
-    pub fn is_list(self) -> bool {
-        self.is_cons() || self.is_nil()
-    }
-
-    /// Iterate over all tails of self.  self should be a list, i.e. a chain
-    /// of cons cells ending in nil.  Otherwise a wrong-type-argument error
-    /// will be signaled.
-    pub fn iter_tails(self) -> TailsIter {
-        TailsIter::new(self, Some(Qlistp))
-    }
-
-    /// Iterate over all tails of self.  If self is not a cons-chain,
-    /// iteration will stop at the first non-cons without signaling.
-    pub fn iter_tails_safe(self) -> TailsIter {
-        TailsIter::new(self, None)
-    }
-
-    /// Iterate over all tails of self.  self should be a plist, i.e. a chain
-    /// of cons cells ending in nil.  Otherwise a wrong-type-argument error
-    /// will be signaled.
-    pub fn iter_tails_plist(self) -> TailsIter {
-        TailsIter::new(self, Some(Qplistp))
-    }
-
-    /// Iterate over the car cells of a list.
-    pub fn iter_cars(self) -> CarIter {
-        CarIter::new(self, Some(Qlistp))
-    }
-
-    /// Iterate over all cars of self. If self is not a cons-chain,
-    /// iteration will stop at the first non-cons without signaling.
-    pub fn iter_cars_safe(self) -> CarIter {
-        CarIter::new(self, None)
-    }
-}
-
-impl From<LispObject> for LispCons {
-    #[inline]
-    fn from(o: LispObject) -> Self {
-        o.as_cons_or_error()
-    }
-}
-
-impl From<LispObject> for Option<LispCons> {
-    #[inline]
-    fn from(o: LispObject) -> Self {
-        if o.is_list() {
-            Some(o.as_cons_or_error())
-        } else {
-            None
-        }
-    }
-}
-
-impl From<LispCons> for LispObject {
-    fn from(c: LispCons) -> Self {
-        c.as_obj()
-    }
-}
-
-/// A newtype for objects we know are conses.
-#[derive(Clone, Copy)]
-pub struct LispCons(LispObject);
-
-impl LispCons {
-    pub fn as_obj(self) -> LispObject {
-        self.0
-    }
-
-    fn _extract(self) -> *mut Lisp_Cons {
-        self.0.get_untaggedptr() as *mut remacs_sys::Lisp_Cons
-    }
-
-    /// Return the car (first cell).
-    pub fn car(self) -> LispObject {
-        unsafe { (*self._extract()).car }
-    }
-
-    /// Return the cdr (second cell).
-    pub fn cdr(self) -> LispObject {
-        unsafe { (*self._extract()).u.cdr }
-    }
-
-    pub fn as_tuple(self) -> (LispObject, LispObject) {
-        (self.car(), self.cdr())
-    }
-
-    /// Set the car of the cons cell.
-    pub fn set_car(self, n: LispObject) {
-        unsafe {
-            (*self._extract()).car = n;
-        }
-    }
-
-    /// Set the car of the cons cell.
-    pub fn set_cdr(self, n: LispObject) {
-        unsafe {
-            (*self._extract()).u.cdr = n;
-        }
-    }
-
-    /// Check that "self" is an impure (i.e. not readonly) cons cell.
-    pub fn check_impure(self) {
-        unsafe {
-            CHECK_IMPURE(self.0, self._extract() as *mut c_void);
-        }
-    }
-}
-
 pub fn is_autoload(function: LispObject) -> bool {
     function
         .as_cons()
         .map_or(false, |cell| cell.car().eq_raw(Qautoload))
 }
 
-// Float support (LispType == Lisp_Float == 7 )
-
-#[test]
-fn test_lisp_float_size() {
-    let double_size = mem::size_of::<EmacsDouble>();
-    let ptr_size = mem::size_of::<*const Lisp_Float>();
-
-    assert!(mem::size_of::<Lisp_Float>() == max(double_size, ptr_size));
-}
-
-pub type LispFloatRef = ExternalPtr<Lisp_Float>;
-
-impl LispFloatRef {
-    pub fn as_data(&self) -> &EmacsDouble {
-        unsafe { &*(&self.u.data as *const EmacsDouble) }
-    }
-}
-
-impl LispObject {
-    #[inline]
-    pub fn is_float(self) -> bool {
-        self.get_type() == Lisp_Type::Lisp_Float
-    }
-
-    #[inline]
-    unsafe fn to_float_unchecked(self) -> LispFloatRef {
-        debug_assert!(self.is_float());
-        LispFloatRef::new(self.get_untaggedptr() as *mut remacs_sys::Lisp_Float)
-    }
-
-    unsafe fn get_float_data_unchecked(self) -> EmacsDouble {
-        *self.to_float_unchecked().as_data()
-    }
-
-    pub fn as_float(self) -> Option<EmacsDouble> {
-        if self.is_float() {
-            Some(unsafe { self.get_float_data_unchecked() })
-        } else {
-            None
-        }
-    }
-
-    pub fn as_float_or_error(self) -> EmacsDouble {
-        if self.is_float() {
-            unsafe { self.get_float_data_unchecked() }
-        } else {
-            wrong_type!(Qfloatp, self)
-        }
-    }
-
-    /*
-    /// If the LispObject is a number (of any kind), get a floating point value for it
-    pub fn any_to_float(self) -> Option<EmacsDouble> {
-        self.as_float()
-            .or_else(|| self.as_fixnum().map(|i| i as EmacsDouble))
-    }
-    */
-
-    pub fn any_to_float_or_error(self) -> EmacsDouble {
-        self.as_float().unwrap_or_else(|| {
-            self.as_fixnum()
-                .unwrap_or_else(|| wrong_type!(Qnumberp, self)) as EmacsDouble
-        })
-    }
-}
-
-impl From<LispObject> for EmacsDouble {
-    #[inline]
-    fn from(o: LispObject) -> Self {
-        o.any_to_float_or_error()
-    }
-}
-
-impl From<LispObject> for Option<EmacsDouble> {
-    #[inline]
-    fn from(o: LispObject) -> Self {
-        if o.is_nil() {
-            None
-        } else {
-            Some(o.any_to_float_or_error())
-        }
-    }
-}
-
-impl From<EmacsDouble> for LispObject {
-    #[inline]
-    fn from(v: EmacsDouble) -> Self {
-        LispObject::from_float(v)
-    }
-}
-
-// String support (LispType == 4)
-
-impl LispObject {
-    #[inline]
-    pub fn is_string(self) -> bool {
-        self.get_type() == Lisp_Type::Lisp_String
-    }
-
-    #[inline]
-    pub fn as_string(self) -> Option<LispStringRef> {
-        if self.is_string() {
-            Some(unsafe { self.as_string_unchecked() })
-        } else {
-            None
-        }
-    }
-
-    #[inline]
-    pub fn as_string_or_error(self) -> LispStringRef {
-        self.as_string()
-            .unwrap_or_else(|| wrong_type!(Qstringp, self))
-    }
-
-    #[inline]
-    pub unsafe fn as_string_unchecked(self) -> LispStringRef {
-        LispStringRef::new(self.get_untaggedptr() as *mut remacs_sys::Lisp_String)
-    }
-
-    #[inline]
-    pub fn empty_unibyte_string() -> LispStringRef {
-        LispStringRef::from(unsafe { empty_unibyte_string })
-    }
-}
-
-impl From<LispObject> for LispStringRef {
-    #[inline]
-    fn from(o: LispObject) -> Self {
-        o.as_string_or_error()
-    }
-}
-
-impl From<LispStringRef> for LispObject {
-    #[inline]
-    fn from(s: LispStringRef) -> Self {
-        s.as_lisp_obj()
-    }
-}
-
 // Other functions
 
-pub trait IsLispNatnum {
-    fn check_natnum(self);
-}
-
-impl IsLispNatnum for EmacsInt {
-    fn check_natnum(self) {
-        if self < 0 {
-            wrong_type!(Qwholenump, LispObject::from(self));
-        }
-    }
-}
-
-#[derive(Clone, Copy)]
-pub enum LispNumber {
-    Fixnum(EmacsInt),
-    Float(f64),
-}
-
-impl LispNumber {
-    pub fn to_fixnum(&self) -> EmacsInt {
-        match *self {
-            LispNumber::Fixnum(v) => v,
-            LispNumber::Float(v) => v as EmacsInt,
-        }
-    }
-}
-
 impl LispObject {
-    #[inline]
-    pub fn is_number(self) -> bool {
-        self.is_fixnum() || self.is_float()
-    }
-
-    /*
-    #[inline]
-    pub fn as_number_or_error(self) -> LispNumber {
-        if let Some(n) = self.as_fixnum() {
-            LispNumber::Fixnum(n)
-        } else if let Some(f) = self.as_float() {
-            LispNumber::Float(f)
-        } else {
-            wrong_type!(Qnumberp, self)
-        }
-    }
-    */
-
-    pub fn as_number_coerce_marker(self) -> Option<LispNumber> {
-        if let Some(n) = self.as_fixnum() {
-            Some(LispNumber::Fixnum(n))
-        } else if let Some(f) = self.as_float() {
-            Some(LispNumber::Float(f))
-        } else if let Some(m) = self.as_marker() {
-            Some(LispNumber::Fixnum(m.charpos_or_error() as EmacsInt))
-        } else {
-            None
-        }
-    }
-
-    #[inline]
-    pub fn as_number_coerce_marker_or_error(self) -> LispNumber {
-        self.as_number_coerce_marker()
-            .unwrap_or_else(|| wrong_type!(Qnumber_or_marker_p, self))
-    }
-
     #[inline]
     pub fn is_nil(self) -> bool {
         self == Qnil
@@ -1567,65 +630,6 @@ impl LispObject {
     #[inline]
     pub fn is_t(self) -> bool {
         self == Qt
-    }
-
-    #[inline]
-    pub fn is_marker(self) -> bool {
-        self.as_misc()
-            .map_or(false, |m| m.get_type() == Lisp_Misc_Type::Lisp_Misc_Marker)
-    }
-
-    #[inline]
-    pub fn as_marker(self) -> Option<LispMarkerRef> {
-        self.as_misc().and_then(|m| {
-            if m.get_type() == Lisp_Misc_Type::Lisp_Misc_Marker {
-                unsafe { Some(mem::transmute(m)) }
-            } else {
-                None
-            }
-        })
-    }
-
-    pub fn as_marker_or_error(self) -> LispMarkerRef {
-        self.as_marker()
-            .unwrap_or_else(|| wrong_type!(Qmarkerp, self))
-    }
-
-    /// Nonzero iff X is a character.
-    pub fn is_character(self) -> bool {
-        self.as_fixnum()
-            .map_or(false, |i| 0 <= i && i <= EmacsInt::from(MAX_CHAR))
-    }
-
-    /// Check if Lisp object is a character or not and return the codepoint
-    /// Similar to CHECK_CHARACTER
-    #[inline]
-    pub fn as_character_or_error(self) -> Codepoint {
-        if !self.is_character() {
-            wrong_type!(Qcharacterp, self)
-        }
-        self.as_fixnum().unwrap() as Codepoint
-    }
-
-    #[inline]
-    pub fn is_overlay(self) -> bool {
-        self.as_misc()
-            .map_or(false, |m| m.get_type() == Lisp_Misc_Type::Lisp_Misc_Overlay)
-    }
-
-    pub fn as_overlay(self) -> Option<LispOverlayRef> {
-        self.as_misc().and_then(|m| {
-            if m.get_type() == Lisp_Misc_Type::Lisp_Misc_Overlay {
-                unsafe { Some(mem::transmute(m)) }
-            } else {
-                None
-            }
-        })
-    }
-
-    pub fn as_overlay_or_error(self) -> LispOverlayRef {
-        self.as_overlay()
-            .unwrap_or_else(|| wrong_type!(Qoverlayp, self))
     }
 
     // The three Emacs Lisp comparison functions.
@@ -1665,72 +669,6 @@ impl LispObject {
 
     pub fn is_function(self) -> bool {
         FUNCTIONP(self)
-    }
-}
-
-impl From<LispObject> for LispNumber {
-    fn from(o: LispObject) -> Self {
-        o.as_number_coerce_marker_or_error()
-    }
-}
-
-impl From<LispObject> for Option<LispNumber> {
-    fn from(o: LispObject) -> Self {
-        o.as_number_coerce_marker()
-    }
-}
-
-impl From<LispObject> for LispMarkerRef {
-    fn from(o: LispObject) -> Self {
-        o.as_marker_or_error()
-    }
-}
-
-impl From<LispMarkerRef> for LispObject {
-    fn from(m: LispMarkerRef) -> Self {
-        m.as_lisp_obj()
-    }
-}
-
-impl From<LispObject> for Option<LispMarkerRef> {
-    fn from(o: LispObject) -> Self {
-        o.as_marker()
-    }
-}
-
-impl From<LispObject> for LispOverlayRef {
-    fn from(o: LispObject) -> Self {
-        o.as_overlay_or_error()
-    }
-}
-
-impl From<LispOverlayRef> for LispObject {
-    fn from(o: LispOverlayRef) -> Self {
-        o.as_lisp_obj()
-    }
-}
-
-impl From<LispObject> for Option<LispOverlayRef> {
-    fn from(o: LispObject) -> Self {
-        o.as_overlay()
-    }
-}
-
-impl From<LispObject> for LispProcessRef {
-    fn from(o: LispObject) -> Self {
-        o.as_process_or_error()
-    }
-}
-
-impl From<LispProcessRef> for LispObject {
-    fn from(p: LispProcessRef) -> Self {
-        p.as_lisp_obj()
-    }
-}
-
-impl From<LispObject> for Option<LispProcessRef> {
-    fn from(o: LispObject) -> Self {
-        o.as_process()
     }
 }
