+++ rust_src/src/windows.rs
@@ -228,6 +228,80 @@ impl LispWindowRef {
     }
 }
 
+impl From<LispObject> for LispWindowRef {
+    fn from(o: LispObject) -> Self {
+        o.as_window_or_error()
+    }
+}
+
+impl From<LispWindowRef> for LispObject {
+    fn from(w: LispWindowRef) -> Self {
+        w.as_lisp_obj()
+    }
+}
+
+impl From<LispObject> for Option<LispWindowRef> {
+    #[inline]
+    fn from(o: LispObject) -> Self {
+        o.as_window()
+    }
+}
+
+impl LispObject {
+    pub fn is_window(self) -> bool {
+        self.as_vectorlike()
+            .map_or(false, |v| v.is_pseudovector(pvec_type::PVEC_WINDOW))
+    }
+
+    pub fn as_window(self) -> Option<LispWindowRef> {
+        self.as_vectorlike().and_then(|v| v.as_window())
+    }
+
+    pub fn as_window_or_error(self) -> LispWindowRef {
+        self.as_window()
+            .unwrap_or_else(|| wrong_type!(Qwindowp, self))
+    }
+
+    pub fn as_minibuffer_or_error(self) -> LispWindowRef {
+        let w = self
+            .as_window()
+            .unwrap_or_else(|| wrong_type!(Qwindowp, self));
+        if !w.is_minibuffer() {
+            error!("Window is not a minibuffer window");
+        }
+        w
+    }
+
+    pub fn as_live_window(self) -> Option<LispWindowRef> {
+        self.as_window()
+            .and_then(|w| if w.is_live() { Some(w) } else { None })
+    }
+
+    pub fn as_live_window_or_error(self) -> LispWindowRef {
+        self.as_live_window()
+            .unwrap_or_else(|| wrong_type!(Qwindow_live_p, self))
+    }
+
+    pub fn as_valid_window(self) -> Option<LispWindowRef> {
+        self.as_window()
+            .and_then(|w| if w.is_valid() { Some(w) } else { None })
+    }
+
+    pub fn as_valid_window_or_error(self) -> LispWindowRef {
+        self.as_valid_window()
+            .unwrap_or_else(|| wrong_type!(Qwindow_valid_p, self))
+    }
+
+    /*
+    pub fn is_window_configuration(self) -> bool {
+        self.as_vectorlike().map_or(
+            false,
+            |v| v.is_pseudovector(pvec_type::PVEC_WINDOW_CONFIGURATION),
+        )
+    }
+    */
+}
+
 pub type LispGlyphMatrixRef = ExternalPtr<glyph_matrix>;
 
 impl LispGlyphMatrixRef {
